+++ expression.h
@@ -81,6 +81,8 @@ class ExpressionC
     bool mergeConstNodes(string & sResult); // merge const expression, reduce calculation during matching. If merged successfully, return true, sResult returns result.
     bool getAggFuncs(unordered_map< string,GroupProp > & aggFuncs); // get the full list of aggregation functions in the expression.
     bool getAnaFuncs(unordered_map< string,vector<ExpressionC> > & anaFuncs, unordered_map< string, vector<int> > & anaGroupNums); // get the full list of analytic functions in the expression.
+    bool getTreeFuncs(unordered_map< string,vector<ExpressionC> > & treeFuncs); // get the full list of hierarchy (tree) functions in the expression.
+    void setTreeFuncs(unordered_map< string,string > & treeFuncVals); // set result for the tree functions
     FunctionC* getAnaFunc(string funcExpStr); // search analytic function in this expression using the function expression string.
     bool calAggFunc(const GroupProp & aggGroupProp, FunctionC* function, string & sResult); // calculate final aggregation function result